+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `SendPolicy` added to API of event-creation for user control of delivery guarantee (reliability and ordering).
+
 ## [0.5.0] - 2023-06-26
 
 ### Added